Rendering bug fixes + visual update & namechange
Fixed two rendering bugs within the GUI:
    1. The 'entry cards' (the boxes with the service name, etc. written in them) now align properly in the rendering panel (before the first row of entries would be misaligned)
    2. The 'add entry'-button (the '+'-button) now does not weirdly resize when resizing the window
Also updated the visuals a bit (namely the colors, but also removed the '?'-icons for the different dialogs). Additionally named the application, is is now called 'PassSafe' (very creative, I know).
The last change I made is making the 'category option'-spinner in the 'add entry'-dialog uneditable, so that the content (and hence the possible categories) cannot be modified (to avoid 'non-existent' categories and resulting problems).